refactor: reorganize homeassistant specific files
