+++ teslajsonpy/connection.py
@@ -16,8 +16,8 @@ from typing import Dict, Text
 import aiohttp
 from yarl import URL
 
-from teslajsonpy.exceptions import IncompleteCredentials, TeslaException
 from teslajsonpy.const import DRIVING_INTERVAL, WEBSOCKET_TIMEOUT
+from teslajsonpy.exceptions import IncompleteCredentials, TeslaException
 
 _LOGGER = logging.getLogger(__name__)
