+++ teslajsonpy/homeassistant/__init__.py
@@ -1,0 +1,6 @@
+"""
+Python Package for controlling Tesla API.
+
+For more details about this api, please refer to the documentation at
+https://github.com/zabuldon/teslajsonpy
+"""
