+++ teslajsonpy/homeassistant/battery_sensor.py
@@ -7,7 +7,7 @@ https://github.com/zabuldon/teslajsonpy
 """
 from typing import Dict, Optional, Text
 
-from teslajsonpy.vehicle import VehicleDevice
+from teslajsonpy.homeassistant.vehicle import VehicleDevice
 
 
 class Battery(VehicleDevice):
