+++ teslajsonpy/homeassistant/vehicle.py
