+++ tests/test_tesla_exception.py
