+++ tests/unit_tests/homeassistant/test_charger_lock.py
@@ -1,11 +1,11 @@
-"""Test door lock."""
+"""Test charger lock."""
 
 import pytest
 
-from tests.tesla_mock import TeslaMock
-
 from teslajsonpy.controller import Controller
-from teslajsonpy.lock import Lock
+from teslajsonpy.homeassistant.lock import ChargerLock
+
+from tests.tesla_mock import TeslaMock
 
 
 def test_has_battery(monkeypatch):
@@ -15,7 +15,7 @@ def test_has_battery(monkeypatch):
     _controller = Controller(None)
 
     _data = _mock.data_request_vehicle()
-    _lock = Lock(_data, _controller)
+    _lock = ChargerLock(_data, _controller)
 
     assert not _lock.has_battery()
 
@@ -27,7 +27,7 @@ def test_is_locked_on_init(monkeypatch):
     _controller = Controller(None)
 
     _data = _mock.data_request_vehicle()
-    _lock = Lock(_data, _controller)
+    _lock = ChargerLock(_data, _controller)
 
     assert not _lock is None
     assert not _lock.is_locked()
@@ -41,8 +41,8 @@ async def test_is_locked_after_update(monkeypatch):
     _controller = Controller(None)
 
     _data = _mock.data_request_vehicle()
-    _data["vehicle_state"]["locked"] = True
-    _lock = Lock(_data, _controller)
+    _data["charge_state"]["charge_port_door_open"] = True
+    _lock = ChargerLock(_data, _controller)
 
     await _lock.async_update()
 
@@ -58,8 +58,8 @@ async def test_lock(monkeypatch):
     _controller = Controller(None)
 
     _data = _mock.data_request_vehicle()
-    _data["vehicle_state"]["locked"] = False
-    _lock = Lock(_data, _controller)
+    _data["charge_state"]["charge_port_door_open"] = False
+    _lock = ChargerLock(_data, _controller)
 
     await _lock.async_update()
     await _lock.lock()
@@ -76,8 +76,8 @@ async def test_lock_already_locked(monkeypatch):
     _controller = Controller(None)
 
     _data = _mock.data_request_vehicle()
-    _data["vehicle_state"]["locked"] = True
-    _lock = Lock(_data, _controller)
+    _data["charge_state"]["charge_port_door_open"] = True
+    _lock = ChargerLock(_data, _controller)
 
     await _lock.async_update()
     await _lock.lock()
@@ -94,8 +94,8 @@ async def test_unlock(monkeypatch):
     _controller = Controller(None)
 
     _data = _mock.data_request_vehicle()
-    _data["vehicle_state"]["locked"] = True
-    _lock = Lock(_data, _controller)
+    _data["charge_state"]["charge_port_door_open"] = True
+    _lock = ChargerLock(_data, _controller)
 
     await _lock.async_update()
     await _lock.unlock()
@@ -112,8 +112,8 @@ async def test_unlock_already_unlocked(monkeypatch):
     _controller = Controller(None)
 
     _data = _mock.data_request_vehicle()
-    _data["vehicle_state"]["locked"] = False
-    _lock = Lock(_data, _controller)
+    _data["charge_state"]["charge_port_door_open"] = False
+    _lock = ChargerLock(_data, _controller)
 
     await _lock.async_update()
     await _lock.unlock()
