+++ tests/unit_tests/homeassistant/test_charger_switch.py
@@ -2,10 +2,10 @@
 
 import pytest
 
-from tests.tesla_mock import TeslaMock
-
 from teslajsonpy.controller import Controller
-from teslajsonpy.charger import ChargerSwitch
+from teslajsonpy.homeassistant.charger import ChargerSwitch
+
+from tests.tesla_mock import TeslaMock
 
 
 def test_has_battery(monkeypatch):
