+++ tests/unit_tests/homeassistant/test_charging_sensor.py
@@ -2,10 +2,10 @@
 
 import pytest
 
-from tests.tesla_mock import TeslaMock
-
 from teslajsonpy.controller import Controller
-from teslajsonpy.charger import ChargingSensor
+from teslajsonpy.homeassistant.charger import ChargingSensor
+
+from tests.tesla_mock import TeslaMock
 
 
 def test_has_battery(monkeypatch):
