+++ tests/unit_tests/homeassistant/test_range_switch.py
@@ -2,10 +2,10 @@
 
 import pytest
 
-from tests.tesla_mock import TeslaMock
-
 from teslajsonpy.controller import Controller
-from teslajsonpy.charger import RangeSwitch
+from teslajsonpy.homeassistant.charger import RangeSwitch
+
+from tests.tesla_mock import TeslaMock
 
 
 def test_has_battery(monkeypatch):
