+++ tests/unit_tests/homeassistant/test_trunk_lock.py
@@ -1,11 +1,11 @@
-"""Test charger lock."""
+"""Test trunk lock."""
 
 import pytest
 
-from tests.tesla_mock import TeslaMock
-
 from teslajsonpy.controller import Controller
-from teslajsonpy.lock import ChargerLock
+from teslajsonpy.homeassistant.trunk import TrunkLock
+
+from tests.tesla_mock import TeslaMock
 
 
 def test_has_battery(monkeypatch):
@@ -15,7 +15,7 @@ def test_has_battery(monkeypatch):
     _controller = Controller(None)
 
     _data = _mock.data_request_vehicle()
-    _lock = ChargerLock(_data, _controller)
+    _lock = TrunkLock(_data, _controller)
 
     assert not _lock.has_battery()
 
@@ -27,9 +27,9 @@ def test_is_locked_on_init(monkeypatch):
     _controller = Controller(None)
 
     _data = _mock.data_request_vehicle()
-    _lock = ChargerLock(_data, _controller)
+    _lock = TrunkLock(_data, _controller)
 
-    assert not _lock is None
+    assert _lock is not None
     assert not _lock.is_locked()
 
 
@@ -41,48 +41,12 @@ async def test_is_locked_after_update(monkeypatch):
     _controller = Controller(None)
 
     _data = _mock.data_request_vehicle()
-    _data["charge_state"]["charge_port_door_open"] = True
-    _lock = ChargerLock(_data, _controller)
+    _data["vehicle_state"]["rt"] = 0
+    _lock = TrunkLock(_data, _controller)
 
     await _lock.async_update()
 
-    assert not _lock is None
-    assert _lock.is_locked()
-
-
-@pytest.mark.asyncio
-async def test_lock(monkeypatch):
-    """Test lock()."""
-
-    _mock = TeslaMock(monkeypatch)
-    _controller = Controller(None)
-
-    _data = _mock.data_request_vehicle()
-    _data["charge_state"]["charge_port_door_open"] = False
-    _lock = ChargerLock(_data, _controller)
-
-    await _lock.async_update()
-    await _lock.lock()
-
-    assert not _lock is None
-    assert _lock.is_locked()
-
-
-@pytest.mark.asyncio
-async def test_lock_already_locked(monkeypatch):
-    """Test lock() when already locked."""
-
-    _mock = TeslaMock(monkeypatch)
-    _controller = Controller(None)
-
-    _data = _mock.data_request_vehicle()
-    _data["charge_state"]["charge_port_door_open"] = True
-    _lock = ChargerLock(_data, _controller)
-
-    await _lock.async_update()
-    await _lock.lock()
-
-    assert not _lock is None
+    assert _lock is not None
     assert _lock.is_locked()
 
 
@@ -94,13 +58,13 @@ async def test_unlock(monkeypatch):
     _controller = Controller(None)
 
     _data = _mock.data_request_vehicle()
-    _data["charge_state"]["charge_port_door_open"] = True
-    _lock = ChargerLock(_data, _controller)
+    _data["vehicle_state"]["rt"] = 0
+    _lock = TrunkLock(_data, _controller)
 
     await _lock.async_update()
     await _lock.unlock()
 
-    assert not _lock is None
+    assert _lock is not None
     assert not _lock.is_locked()
 
 
@@ -112,11 +76,53 @@ async def test_unlock_already_unlocked(monkeypatch):
     _controller = Controller(None)
 
     _data = _mock.data_request_vehicle()
-    _data["charge_state"]["charge_port_door_open"] = False
-    _lock = ChargerLock(_data, _controller)
+    _data["vehicle_state"]["rt"] = 123
+    _lock = TrunkLock(_data, _controller)
 
     await _lock.async_update()
     await _lock.unlock()
 
-    assert not _lock is None
+    assert _lock is not None
     assert not _lock.is_locked()
+
+    # Reset to default for next tests
+    _data["vehicle_state"]["rt"] = 0
+
+
+@pytest.mark.asyncio
+async def test_lock(monkeypatch):
+    """Test lock()."""
+
+    _mock = TeslaMock(monkeypatch)
+    _controller = Controller(None)
+
+    _data = _mock.data_request_vehicle()
+    _data["vehicle_state"]["rt"] = 123
+    _lock = TrunkLock(_data, _controller)
+
+    await _lock.async_update()
+    await _lock.lock()
+
+    assert _lock is not None
+    assert _lock.is_locked()
+
+    # Reset to default for next tests
+    _data["vehicle_state"]["rt"] = 0
+
+
+@pytest.mark.asyncio
+async def test_lock_already_locked(monkeypatch):
+    """Test lock() when already locked."""
+
+    _mock = TeslaMock(monkeypatch)
+    _controller = Controller(None)
+
+    _data = _mock.data_request_vehicle()
+    _data["vehicle_state"]["rt"] = 0
+    _lock = TrunkLock(_data, _controller)
+
+    await _lock.async_update()
+    await _lock.lock()
+
+    assert _lock is not None
+    assert _lock.is_locked()
